完成 AndroidJar 功能

+++ src/main/groovy/com/mingyuan/gradle/PackageAndroidJarTask.groovy
@@ -1,11 +1,14 @@
-package com.mingyuan.gradle;
+package com.mingyuan.gradle
 
-import com.mingyuan.gradle.exception.PackageException;
-import org.gradle.api.Project;
+import com.mingyuan.gradle.exception.PackageException
+import org.gradle.api.Project
 import org.gradle.api.java.archives.Attributes
-import org.gradle.api.tasks.ParallelizableTask;
-import org.gradle.api.tasks.TaskAction;
-import org.gradle.api.tasks.bundling.Jar;
+import org.gradle.api.tasks.ParallelizableTask
+import org.gradle.api.tasks.TaskAction
+import org.gradle.api.tasks.bundling.Jar
+
+import java.util.zip.ZipEntry
+import java.util.zip.ZipFile
 
 /**
  * Created by yanxq on 17/1/10.
@@ -18,16 +21,15 @@ public class PackageAndroidJarTask extends Jar {
     @TaskAction
     public void onTaskAction() {
         Project target = getProject();
-
-        prepareManifest();
-
         PackageExtension extension = (PackageExtension) target.getExtensions()
                 .getByName("packageExtension");
+
+        prepareManifest(extension);
 
         File archiveFile = extension.getArchiveFile(getProject());
         deleteOldArchiveFileIfExist(archiveFile);
 
-        File classesDirFile = new File(target.getProjectDir(),"build/intermediates/classes/release/com/");
+        File classesDirFile = new File(target.getProjectDir(),"build/intermediates/classes/release/");
         if (!classesDirFile.exists()) {
             throw new PackageException(String.format("%s is not exist!", classesDirFile.getPath()));
         }
@@ -36,32 +38,80 @@ public class PackageAndroidJarTask extends Jar {
         for (String excludeFile : EXCLUDE_FILES) {
             exclude(excludeFile);
         }
+        settingMergeFiles(target,extension);
 
         setBaseName(extension.jarBaseName);
         setVersion(extension.jarVersion);
+        setExtension("jar");
+        setDestinationDir(new File(target.getProjectDir(),extension.getArchiveFileDir()))
 
-        target.println("archive file path :  " + new File(target.getProjectDir(),extension.getArchiveFileDir()).getAbsolutePath())
-        into(new File(target.getProjectDir(),extension.getArchiveFileDir()).getAbsolutePath());
-
-        super.copy();
-
-
+        startBuild();
     }
 
+    private void settingMergeFiles(Project target ,PackageExtension extension) {
+        if (extension.mergeJarFiles == null || extension.mergeJarFiles.length == 0) {
+            return;
+        }
 
-    private void prepareManifest() {
+        HashMap<String,String> dependencies = new HashMap<String,String>();
+        for (File dependence : target.configurations.getByName("compile").getFiles()) {
+            dependencies.put(dependence.getName(),dependence);
+        }
+
+        if (extension.mergeJarFiles != null && extension.mergeJarFiles.length > 0) {
+            File projectDependence;
+            for (String mergeFileName : extension.mergeJarFiles) {
+                if ((projectDependence = dependencies.get(mergeFileName)) != null) {
+                    handleManifestProperties(projectDependence)
+                    from(target.zipTree(projectDependence.getAbsolutePath()))
+                } else {
+                    throw new PackageException("Couldn't find merge file : " + mergeFileName);
+                }
+            }
+        }
+    }
+
+    private void handleManifestProperties(File dependence) {
+        ZipFile zipFile = new ZipFile(dependence);
+        ZipEntry metaEntry = zipFile.getEntry("META-INF/MANIFEST.MF");
+        InputStream inputStream = zipFile.getInputStream(metaEntry);
+        Properties properties = new Properties();
+        properties.load(inputStream);
+
+        Attributes attributes = getManifest().getAttributes();
+        for (String proper : properties.stringPropertyNames()) {
+            if (attributes.get(proper) == null) {
+                attributes.put(proper,properties.getProperty(proper));
+            }
+        }
+    }
+
+    private void startBuild() {
+        copy();
+    }
+
+    private void prepareManifest(PackageExtension extension) {
         Attributes attributes = getManifest().getAttributes();
         attributes.put("git-SHA1", TaskUtils.getGitSHA1Value());
         attributes.put("build-date", TaskUtils.getBuildDate());
-        // TODO: 17/1/10 待补充
+
+        if (extension.metaProperties != null && !extension.metaProperties.isEmpty()) {
+            for (Map.Entry<String,String> entry : extension.metaProperties) {
+                attributes.put(entry.getKey(),entry.getValue());
+            }
+        }
     }
 
     private void deleteOldArchiveFileIfExist(File archiveFile) {
         File archiveDirFile = archiveFile.getParentFile();
         if (archiveDirFile.exists()) {
             for (File child : archiveDirFile.listFiles()) {
-                child.delete();
+                if (child.isFile()) {
+                    child.delete();
+                }
             }
+        } else {
+            archiveDirFile.mkdirs();
         }
     }
